add: animus 3 features ported to animus 2.7

+++ animus-2/keyboard-specific-builds/terminus-mini/animus/millis.ino
@@ -6,7 +6,8 @@ boolean readyMillis = false;
 void MillisLoop()
 {
   currentMillis = millis();
-  if ((unsigned long)(currentMillis - previousMillis) >= MILLIS_TICK)
+
+  if(currentMillis - previousMillis >= MILLIS_TICK) // this elapses every 1-1.5 ms
   {
     previousMillis = currentMillis;
     readyMillis = true;
